\name{ASCAdata}
\alias{ASCAdata}
\docType{data}
\title{
Example data for ASCA.Calculate
}
\description{
The example contains data for a two factor experimental design. The first factor
has two treatment levels, the second has three treatment levels. The design is balanced with 10 observations per factor/treatment level. 
}
\usage{data(ASCAdata)}
\format{
  The format is:  \cr
  \link{ASCAX} numerical matrix of dimension 60 4\cr
  \link{ASCAF} numerical matrix of dimension 60 2
}
\details{
The
}

\keyword{ASCA}
