\name{ASCAF}
\alias{ASCAF}
\docType{data}
\title{
Indicator matrix for ASCA example.
}
\description{
The matrix ASCAF gives the factor/treatment combinations for the rows of the data matrix \link{ASCAX}. 
}
\usage{data(ASCAdata)}
\format{
  The format is: \cr
  numerical matrix of dimension 60 2
}
\details{
Each column in ASCAF represents a factor, the numbers in the columns correspond to the treatment level of that factor. In this example, the first factor has two treatment levels, indicated by 1 and 2 in the first column of ASCAF. The second factor has three treatment levels, indicated by the numbers 1, 2 and 3 in the second column. As an example, the two numbers 1 1 in the first row of ASCAF indicate that the first measurement was made on a sample in treatment level 1 for both factors.
}

\keyword{ASCA}
