\name{ASCAX}
\alias{ASCAX}
\docType{data}
\title{
Data matrix for ASCA example
}
\description{
The data matrix ASCAX contains the measurements of 4 variables in a two factor experimental design. 
}
\usage{data(ASCAdata)}
\format{
  The format is: \cr
  numerical matrix, dim(ASCAX) = 60 4
}
\details{
The rows of the data matrix ASCAX contain measurements of 4 variables for an 
experimental design with two experimental factors. The first factor has 2 treatment
levels, the second factor has 3 treatment levels. The assignments of the rows to the 
factor/treatment combinations is done trhough the matrix \link{ASCAF}. The design is 
balanced with 10 measurements per factor/treatment combination.
}

\keyword{ASCA}
