\encoding{latin1}
\name{MetStaT.ConcatWithStringPars}
\alias{MetStaT.ConcatWithStringPars}
\title{
Concatenates an array of strings
}
\description{
Takes an array of string values and simply pastes them together to create one uninterrupted string value.
}
\usage{
MetStaT.ConcatWithStringPars(string)
}

\arguments{
  \item{string}{the array containing the string values.
  }
}

\value{A single string value.
}

\author{
Tim Dorscheidt
}

\keyword{String }
\keyword{Concatenate }
