\name{MetStaT.KillAllDevices}
\alias{MetStaT.KillAllDevices}

\title{
Kills all currently active plot windows.
}

\description{All active plot windows are shut down.}

\usage{
MetStaT.KillAllDevices()
}

\value{
returns nothing, except for the result of all plotting devices being shut down.
}

\author{
Tim Dorscheidt
}

\keyword{ Matrix }
\keyword{ File }