\name{MetStaT.mldivide}
\alias{MetStaT.mldivide}

\title{
Matrix left hand division using a copy of the 'pracma' package 'mldivide' method.
}
\description{
This is a verbatim copy of the 'mldivide' function contained in the 'pracma' package by Hans W. Borchers (under GPL>=3 license), as found on CRAN on November 27, 2012. Together with the 'mrdivide' method, these are the only two 'pracma' package methods used in the 'MetStaT' package, which is why they were copied instead of depending on the entire 'pracma' package. For details on the mldivide method, please see the help file in the 'pracma' package.   
}
\usage{
  MetStaT.mldivide(A, B)
}

\arguments{
  \item{A}{
  }
  \item{B}{
  }
}

\author{
Original author: Hans W. Borchers. Copied and included in the MetStaT package by Tim Dorscheidt.
}
