\name{QStatX}
\alias{QStatX}
\docType{data}
\title{
Data matrix for QStat.Calculate example
}
\description{
This provides an example data matrix for the QStat.Calcualte method
}
\usage{data(QStat)}
\format{
  The format is:
  numeric matrix dim(QStatX) = 18 6
}
\details{
The matrix contains 18 rows, each row represents measurements of 6 variables.
The first 10 rows are normally distributed numbers with zero mean and unit variance.
The remaining 8 rows each have normally distributed numbers with mean 1 and variance 2.
}
\source{
  Global test for metabolic pathway differences between conditions. Diana M. Hendrickx, Huub C.J. Hoefsloot, Margriet M.W.B. Hendriks, Andre. Canelas and Age K. Smilde. Analytica Chimica Acta, Volume 719 (2012), pages 8 - 15
}

\keyword{QStat.Calculate}
