\name{QStatY}
\alias{QStatY}
\docType{data}
\title{
Matrix for QStat.Calculate example
}
\description{
This provides the indicator matrix for the QStat.Calcualte method
}
\usage{data(QStat)}
\format{
  The format is:
  numeric matrix dim(QStatX) = 18 1
}
\details{
The matrix contains 18 rows, each row indicates whether the outcome belongs to the 
first condition, 0, or two the second condition, 1.
}
\source{
  Global test for metabolic pathway differences between conditions. Diana M. Hendrickx, Huub C.J. Hoefsloot, Margriet M.W.B. Hendriks, Andre. Canelas and Age K. Smilde. Analytica Chimica Acta, Volume 719 (2012), pages 8 - 15
}

\keyword{QStat.Calculate}
