\name{SteadyState}
\alias{SteadyState}
\docType{data}
\title{
Example data for RevNetJacobianMethod
}
\description{
Steady state concentrations used by the Jacobian method to reverse engineer a metabolomic network. 
}

\usage{data(RevNetJacobian)}
\format{
    SteadyState, numerical matrix with steady state concentrations\cr
    dim(SteadyState) = 1 4
 }

\source{
Chassagnole, C., D. A. Fell, et al. (2001). "Control of the threonine-synthesis pathway in Escherichia coli: a theoretical and experimental approach." Biochemical Journal 356: 433-444.
}
\references{
Reverse engineering of metabolic networks, a critical assessment. Diana M. Hendrickx, Margriet M. W. B. Hendriks, Paul H. C. Eilers, Age K. Smilde and Huub C. J. Hoefsloot. Mol. BioSyst, Volume 7:2 (2011) pages 511-520
}

\keyword{RevNetJacobianMethod}
\keyword{SteadyState}
